Update PayPerPlay.sol
Removed the received() fallback function for accepting ETH payments.  It is not required and the accounting becomes confusing because of the mixed $MUSIC / ETH currencies if it is accepted.  Distribution would then have to work with both ETH and $MUSIC when in fact this contract should only be receiving $MUSIC