Create balance-migration-error-fix.js
If the migration stops at any point due to failure or manual exiting, it is possible a transfer was completed without updating the database.  Restarting the migration will not allow for a duplicate migration because the DB flags will indicate a transfer was in progress, however, if no transfer was made then restarting the migration won't result in those accounts getting funds.  This script checks the DB and the chain for any incomplete records and fixes the DB state.  Once done, restart migration.  DO NOT RUN THIS WITH THE MIGRATION RUNNING AT THE SAME TIME.